requirements.txt
mcp>=1.0.0
mysql-connector-python>=9.1.0
pypinyin>=0.48.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
starlette>=0.27.0